# C#/AterWebTemplateUsing/src/Http.API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AterWebTemplateUsing",
    "description": "API 文档",
    "version": "v1"
  },
  "paths": {
...
    "/api/SystemUser/filter": {
      "post": {
        "tags": [
          "SystemUser"
        ],
        "summary": "筛选",
        "operationId": "SystemUser_Filter",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemUserFilterDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemUserFilterDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SystemUserFilterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUserItemDtoPageList"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUserItemDtoPageList"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUserItemDtoPageList"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/SystemUser/{id}": {
      "put": {
        "tags": [
          "SystemUser"
        ],
        "summary": "更新",
        "operationId": "SystemUser_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemUserUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemUserUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SystemUserUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUser"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUser"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "SystemUser"
        ],
        "summary": "详情",
        "operationId": "SystemUser_GetDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUser"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUser"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "SystemUser"
        ],
        "summary": "⚠删除",
        "operationId": "SystemUser_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUser"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUser"
                }
              }
            }
          }
        }
      }
    },
    "/api/SystemUser/upload": {
      "post": {
        "tags": [
          "SystemUser"
        ],
        "summary": "上传图片",
        "operationId": "SystemUser_UploadImg",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResult"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Sex": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "description": "性别",
        "format": "int32",
        "x-enumNames": [
          "Male",
          "Female",
          "Else"
        ]
      },
      "SystemRole": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          },
          "updatedTime": {
            "type": "string",
            "format": "date-time"
          },
          "isDeleted": {
            "type": "boolean",
            "description": "软删除"
          },
          "name": {
            "maxLength": 30,
            "type": "string",
            "description": "角色显示名称",
            "nullable": true
          },
          "nameValue": {
            "type": "string",
            "description": "角色名，系统标识",
            "nullable": true
          },
          "isSystem": {
            "type": "boolean",
            "description": "是否系统内置,系统内置不可删除"
          },
          "icon": {
            "maxLength": 30,
            "type": "string",
            "description": "图标",
            "nullable": true
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SystemUser"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "角色表"
      },
...
      "SystemUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          },
          "updatedTime": {
            "type": "string",
            "format": "date-time"
          },
          "isDeleted": {
            "type": "boolean",
            "description": "软删除"
          },
          "userName": {
            "maxLength": 30,
            "type": "string",
            "description": "用户名",
            "nullable": true
          },
          "realName": {
            "maxLength": 30,
            "type": "string",
            "description": "真实姓名",
            "nullable": true
          },
          "email": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "boolean"
          },
          "passwordHash": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          },
          "passwordSalt": {
            "maxLength": 60,
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "phoneNumberConfirmed": {
            "type": "boolean"
          },
          "twoFactorEnabled": {
            "type": "boolean"
          },
          "lockoutEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lockoutEnabled": {
            "type": "boolean"
          },
          "accessFailedCount": {
            "type": "integer",
            "format": "int32"
          },
          "lastLoginTime": {
            "type": "string",
            "description": "最后登录时间",
            "format": "date-time",
            "nullable": true
          },
          "retryCount": {
            "type": "integer",
            "description": "密码重试次数",
            "format": "int32"
          },
          "avatar": {
            "maxLength": 200,
            "type": "string",
            "description": "头像url",
            "nullable": true
          },
          "systemRoles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SystemRole"
            },
            "nullable": true
          },
          "sex": {
            "$ref": "#/components/schemas/Sex"
          }
        },
        "additionalProperties": false,
        "description": "系统用户"
      },
...
      "SystemUserFilterDto": {
        "type": "object",
        "properties": {
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "orderBy": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean",
              "nullable": true
            },
            "description": "排序",
            "nullable": true
          },
          "userName": {
            "maxLength": 30,
            "type": "string",
            "description": "用户名",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "boolean",
            "nullable": true
          },
          "phoneNumberConfirmed": {
            "type": "boolean",
            "nullable": true
          },
          "twoFactorEnabled": {
            "type": "boolean",
            "nullable": true
          },
          "lockoutEnabled": {
            "type": "boolean",
            "nullable": true
          },
          "accessFailedCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "retryCount": {
            "type": "integer",
            "description": "密码重试次数",
            "format": "int32",
            "nullable": true
          },
          "sex": {
            "$ref": "#/components/schemas/Sex"
          }
        },
        "additionalProperties": false,
        "description": "系统用户查询筛选"
      },
      "SystemUserItemDto": {
        "type": "object",
        "properties": {
          "userName": {
            "maxLength": 30,
            "type": "string",
            "description": "用户名",
            "nullable": true
          },
          "realName": {
            "maxLength": 30,
            "type": "string",
            "description": "真实姓名",
            "nullable": true
          },
          "email": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "boolean"
          },
          "phoneNumber": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "phoneNumberConfirmed": {
            "type": "boolean"
          },
          "twoFactorEnabled": {
            "type": "boolean"
          },
          "lockoutEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lockoutEnabled": {
            "type": "boolean"
          },
          "accessFailedCount": {
            "type": "integer",
            "format": "int32"
          },
          "lastLoginTime": {
            "type": "string",
            "description": "最后登录时间",
            "format": "date-time",
            "nullable": true
          },
          "retryCount": {
            "type": "integer",
            "description": "密码重试次数",
            "format": "int32"
          },
          "avatar": {
            "maxLength": 200,
            "type": "string",
            "description": "头像url",
            "nullable": true
          },
          "sex": {
            "$ref": "#/components/schemas/Sex"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          },
          "updatedTime": {
            "type": "string",
            "format": "date-time"
          },
          "isDeleted": {
            "type": "boolean",
            "description": "软删除"
          }
        },
        "additionalProperties": false,
        "description": "系统用户列表元素"
      },
      "SystemUserItemDtoPageList": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SystemUserItemDto"
            },
            "nullable": true
          },
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "SystemUserUpdateDto": {
        "type": "object",
        "properties": {
          "userName": {
            "maxLength": 30,
            "type": "string",
            "description": "用户名",
            "nullable": true
          },
          "realName": {
            "maxLength": 30,
            "type": "string",
            "description": "真实姓名",
            "nullable": true
          },
          "email": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "boolean",
            "nullable": true
          },
          "phoneNumber": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "phoneNumberConfirmed": {
            "type": "boolean",
            "nullable": true
          },
          "twoFactorEnabled": {
            "type": "boolean",
            "nullable": true
          },
          "lockoutEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lockoutEnabled": {
            "type": "boolean",
            "nullable": true
          },
          "accessFailedCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lastLoginTime": {
            "type": "string",
            "description": "最后登录时间",
            "format": "date-time",
            "nullable": true
          },
          "retryCount": {
            "type": "integer",
            "description": "密码重试次数",
            "format": "int32",
            "nullable": true
          },
          "avatar": {
            "maxLength": 200,
            "type": "string",
            "description": "头像url",
            "nullable": true
          },
          "sex": {
            "$ref": "#/components/schemas/Sex"
          }
        },
        "additionalProperties": false,
        "description": "系统用户更新时请求结构"
      },
      "UploadResult": {
        "type": "object",
        "properties": {
          "filePath": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter a valid token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}